requirements.txt
streamlit 
streamlit-carousel
python-dotenv
aiohttp #This is for telegram
requests
asyncio
gradio_client
deep_translator
groq
pillow
webbrowser



